requirements.txt
python-dotenv==1.0.1
streamlit==1.35.0
streamlit_option_menu==0.3.13
google-generativeai==0.8.2